requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
python-multipart
playwright
playwright-stealth
openai
pytesseract
Pillow
yt-dlp
pytest
pytest-asyncio
